third_party/WebKit/LayoutTests/platform/win/svg/animations/svglength-animation-px-to-exs-expected.txt

SVG 1.1 dynamic animation tests
Test SVGLength animation from px to exs.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS rect.width.animVal.value is 100
PASS rect.width.baseVal.value is 100
FAIL rect.width.animVal.value should be close to 150. Was 130.
PASS rect.width.baseVal.value is 100
FAIL rect.width.animVal.value should be close to 200. Was 159.98500061035156.
PASS rect.width.baseVal.value is 100
PASS rect.width.animVal.value is 100
PASS rect.width.baseVal.value is 100
PASS successfullyParsed is true

TEST COMPLETE
